Concurrency in parallel runs. (#927)
Use a concurrent dictionary for managing the proxy managers and handlers
for parallel runs. It is possible that the child runs end up at the same
time. It is necessary for ParallelProxyOperationManager to be thread safe.

Remove MemoryTraceWriter for data serializer since it is not thread safe.

Enabled the parallel integration tests.