Refactored console runner.
Moved the console runner logic to a library. The actual application
will only invoke the business logic from the library.

The console runner app config provides a private probing path for the
library. This will allow us to keep the runner app in a different location
than the console runner library.