Fix writing to trx when error has no message
Fixes intialization of error info that would require errors to have error message and not just stack trace. This is triggered for example by Assert.Fail() in nUnit that provides a null error message. Removed the constructor that we don't need, and enabled initialization in both setters to avoid failing.

Fixes #2319