Enable TreatWarningsAsErrors only on CI
We raised the severity of some rules from message to warning to ensure they are followed in the team but because of the `TreatWarningsAsErrors` we end up with some "boring" dev cycles where we are forced to cleanup usings etc while we iterate. The idea is to keep warnings locally to not slow us down but still fail in PRs so that we can keep styling consistency.