Update devcontainer.json
Removed commented extensions: Removed commented extensions to clean up the code and improve readability.

Added "builtBy" field: Added a new field called "builtBy" to indicate the entity or organization responsible for building this configuration. In this case, it mentions "Pipeline Foundation". It provides credit and recognition to the appropriate entity.

By improving the formatting, removing unnecessary comments, and adding the "builtBy" field, the code becomes more concise and easier to read.