Create a proper MSBuild ToolTask based VSTestTask (#2702)
* Rename VSTestTask to VSTestForwardingTask (#680)

VSTestForwardingTask redirects the console output instead of using MSBuild logs. Renaming it will allow to create a new VSTestTask that uses the MSBuild log infrastructure.

* Refactor the VSTestForwardingTask to use the MSBuild infrastructure (#680)

* Create a ToolTask based VSTestTask that integrates nicely in the MSBuild infrastructure (#680)

This new task will be used by default during the build. If the $(VSTestUseConsole) property is set to True, the old console forwarding VSTestForwardTask will be used instead.

* Initialize the VsTestUseConsole property during acceptance tests

The `$(VSTestUseConsole)` should be set to `True` when running the `dotnet test` command to keep the legacy behavior. Adding it during the tests for now: it should be set in the command itself eventually.

* Fix various issues raised by the latest review

* Simplify the handling of null values

* Log messages to MSBuild instead of to the console

* Integrate changes from the latest code review

* Patch dotnet with newest build task

* wip

* rnm

* revert tests

* tests

* Try adding msbuild specialized logger to output errors on tests

* Use simpler logger

* wip

* fix no-source

* fix no-source

* add message for full error

* Revert playground

* Revert console logger

* Revert integration tests

* Revert change in integration tests

* Fix test

* Revert test unignore

---------

Co-authored-by: Jakub Jareš <me@jakubjares.com>