Set NetCurrent so that it doesn't roll forward automatically
The VSTest repo can't use a floating NetCurrent value as the NetCurrent TFM is actually hardcoded in a number of places => .nuspec files. Set it to net9.0 explicitly so that the build continues to work with a .NET 10 SDK and Arcade 10 SDK (which is used inside the VMR).

Don't hardcode the value when building from source as in that configuration, the TFM needs to actually target the very latest.