Condition property on .NET FX MSBuild
The `BuildWithNetFrameworkHostedCompiler` property forces a .NET
Framework specific compiler package to be used. This property was being
unconditionally set which meant it loaded inside of `dotnet build`. In
that environment the build task is invalid / unsupported.

In this particular case the task ended up throwing errors trying to
connect to the compiler server. That is an expected failure path in the
build task and it will fallback silently to csc.exe. That meant this
went unnoticed as builds remained functional, just significantly slower.

Related: https://github.com/dotnet/sdk/pull/48557