Remove stale copy of S.ComponentModel.Composition from testplatform packages (#15287)
* Remove stale copy of S.ComponentModel.Composition from testplatform packages

S.ComponentModel.Composition is no longer used in testplatform packages, so we can remove the stale copy of it. The assembly that was redistributed targets netstandard2.0 and is a PNSE - throws PlatformNotSupportedException. This proves that the assembly never got loaded and is indeed unused in all cases.

This resolves having this stale copy in the SDK layout under i.e. "C:\Program Files\dotnet\sdk\10.0.100-rc.1.25451.107\System.ComponentModel.Composition.dll"

* Update Version.Details.props

* Fix expected file counts for TestPlatform packages